split middleware
