+++ src/components/HybridApp/HybridAppOnboarding/index.ios.tsx
@@ -1,0 +1,45 @@
+import {useEffect} from 'react';
+import {NativeModules} from 'react-native';
+import type {OnyxEntry} from 'react-native-onyx';
+import Log from '@libs/Log';
+import ONYXKEYS from '@src/ONYXKEYS';
+import type {TryNewDot} from '@src/types/onyx';
+import {useOnyx} from '../../../../__mocks__/react-native-onyx';
+
+const onboardingStatusSelector = (tryNewDot: OnyxEntry<TryNewDot>) => {
+    let completedHybridAppOnboarding = tryNewDot?.classicRedirect?.completedHybridAppOnboarding;
+
+    if (typeof completedHybridAppOnboarding === 'string') {
+        completedHybridAppOnboarding = completedHybridAppOnboarding === 'true';
+    }
+
+    return completedHybridAppOnboarding;
+};
+
+function HybridAppOnboarding() {
+    const [completedHybridAppOnboarding] = useOnyx(ONYXKEYS.NVP_TRYNEWDOT, {selector: onboardingStatusSelector});
+
+    /**
+     * This useEffect tracks changes of `nvp_tryNewDot` value.
+     * We propagate it from OldDot to NewDot with native method due to limitations of old app.
+     */
+    useEffect(() => {
+        if (completedHybridAppOnboarding === undefined) {
+            return;
+        }
+
+        if (!NativeModules.HybridAppModule) {
+            Log.hmmm(`[HybridApp] Onboarding status has changed, but the HybridAppModule is not defined`);
+            return;
+        }
+
+        Log.info(`[HybridApp] Onboarding status has changed. Propagating new value to OldDot`, true, {completedHybridAppOnboarding});
+        NativeModules.HybridAppModule.completeOnboarding(completedHybridAppOnboarding);
+    }, [completedHybridAppOnboarding]);
+
+    return null;
+}
+
+HybridAppOnboarding.displayName = 'HybridAppOnboarding';
+
+export default HybridAppOnboarding;
